backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
av==11.0.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
distro==1.9.0
docopt==0.6.2
eval_type_backport==0.2.2
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.0
jsonpath-python==1.0.6
kombu==5.4.2
llvmlite==0.44.0
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
mistralai==1.5.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.62.0
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
yarl==1.19.0
msal==1.26.0
requests==2.32.3
python-jose==3.3.0
aiosqlite==0.21.0